Refactored thruster arduino libraries to have only MultiplexedBasicESC
